audiopolicy: engineconfigurable: Don't include the example
When using the configurable policy engine, the audiopolicy makefile
implicitly declares dependencies on the configuration files. However,
the example folder already contains the example files which we wan't to
override.

This patch removes the example from the makefile inclusion path so that
we can declare it on a per-device basis without getting errored out by
duplicate module names.

Change-Id: Ie47ebff6b13606c93630c017252ebaa204203a76
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>